third_party/WebKit/LayoutTests/fast/dom/StyleSheet/detached-stylesheet-without-wrapper.html

<html>
<head>
<style>
#test-element {
    text-decoration: line-through;
}
</style>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<span id="test-element"></div>
<script>
description("Test that detached CSSOM subtrees keeps a reference to CSS stylesheets lacking wrappers.");

var rule = document.styleSheets[0].rules.item(0);
document.documentElement.removeChild(document.head);
shouldNotBe("rule.parentStyleSheet", "null");
</script>
</body>
</html>
